+++ lib/screens/chat/chat_screen/image_screen.dart
@@ -1,0 +1,11 @@
+// import 'package:flutter/cupertino.dart';
+
+// class ImageScreen extends StatelessWidget {
+//   Uri
+//   const ImageScreen({super.key});
+
+//   @override
+//   Widget build(BuildContext context) {
+//     return const Placeholder();
+//   }
+// }
